Simplify the return from GetWallclock.  LSL_Floats are internally doubles so no need to cast to float.

+++ OpenSim/Region/ScriptEngine/Shared/Api/Implementation/LSL_Api.cs
@@ -3115,7 +3115,7 @@ namespace OpenSim.Region.ScriptEngine.Shared.Api
         {
             var tzinfo = TimeZoneInfo.FindSystemTimeZoneById("Pacific Standard Time");
             var dateTime = TimeZoneInfo.ConvertTime(DateTimeOffset.UtcNow, tzinfo).DateTime;
-            return (float)Math.Truncate(dateTime.TimeOfDay.TotalSeconds);
+            return Math.Truncate(dateTime.TimeOfDay.TotalSeconds);
         }
 
         public LSL_Float llGetTime()
